test_case_generator/requirements.txt
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
deap==1.4.3
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
MarkupSafe==3.0.2
numpy==2.0.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
pathable==0.4.4
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rpds-py==0.26.0
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
